rename to server-config
